APIs consertadas (Java)

+++ java/src/main/java/com/backendpi/backend/repository/PostSalvoRepository.java
@@ -3,5 +3,5 @@ package com.backendpi.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.backendpi.backend.model.PostSalvo;
 
-public interface PostSalvoRepository extends JpaRepository<PostSalvo, Integer> {
+public interface PostSalvoRepository extends JpaRepository<PostSalvo, Long> {
 }
