+++ java/src/main/java/com/backendpi/backend/repository/VotoRepository.java
@@ -3,5 +3,5 @@ package com.backendpi.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.backendpi.backend.model.Voto;
 
-public interface VotoRepository extends JpaRepository<Voto, Integer> {
+public interface VotoRepository extends JpaRepository<Voto, Long> {
 }
